refactor: extract duplicated handleValidationError into shared module (#216)
The same validation-error handler was copy-pasted across four command
files (console.ts:47, peek.ts:43, tail.ts:28, network/list.ts:102).
Extract to src/commands/shared/handleValidationError.ts and import.

Net: -98 lines, one source of truth for validation error presentation.
