+++ src/commands/peek.ts
@@ -13,16 +13,14 @@ import {
   type FetchResult,
 } from '@/commands/shared/dataFetcher.js';
 import { setupFollowMode } from '@/commands/shared/followMode.js';
+import { handleValidationError } from '@/commands/shared/handleValidationError.js';
 import type { PeekCommandOptions } from '@/commands/shared/optionTypes.js';
 import { positiveIntRule, resourceTypeRule } from '@/commands/shared/validation.js';
 import type { Protocol } from '@/connection/typed-cdp.js';
 import { filterByResourceType } from '@/telemetry/filters.js';
 import type { BdgOutput } from '@/types.js';
-import { OutputBuilder } from '@/ui/OutputBuilder.js';
-import { CommandError } from '@/ui/errors/index.js';
 import { formatPreview, type PreviewOptions } from '@/ui/formatters/preview.js';
 import { followingPreviewMessage, stoppedFollowingPreviewMessage } from '@/ui/messages/preview.js';
-import { EXIT_CODES } from '@/utils/exitCodes.js';
 
 interface ProcessedPreview {
   output: BdgOutput;
@@ -38,26 +36,6 @@ function parseOptions(options: PeekCommandOptions): ParsedOptions {
   const lastN = positiveIntRule({ min: 1, max: 1000, default: 10 }).validate(options.last);
   const resourceTypes = resourceTypeRule().validate(options.type);
   return { lastN, resourceTypes };
-}
-
-function handleValidationError(error: unknown, json: boolean): never {
-  if (error instanceof CommandError) {
-    if (json) {
-      const errorOptions: { exitCode: number; suggestion?: string } = {
-        exitCode: error.exitCode,
-      };
-      if (error.metadata.suggestion) {
-        errorOptions.suggestion = error.metadata.suggestion;
-      }
-      console.log(JSON.stringify(OutputBuilder.buildJsonError(error.message, errorOptions)));
-    } else {
-      console.error(error.message);
-      if (error.metadata.suggestion) console.error(error.metadata.suggestion);
-    }
-    process.exit(error.exitCode);
-  }
-  console.error(error instanceof Error ? error.message : String(error));
-  process.exit(EXIT_CODES.INVALID_ARGUMENTS);
 }
 
 async function fetchAndFilterPreview(
